New synthesizer. Cleanup.
